# apps/worker/openapi-restful.yaml
openapi: 3.0.3
info:
  title: Sutradhar RESTful Actions API
  description: |
    RESTful API endpoints for GitHub, Slack, and Google Calendar operations.
    
    All endpoints follow REST principles with resource-based URLs, proper HTTP methods,
    and consistent response formats.
  version: 1.0.0
servers:
  - url: http://localhost:2198/api/v1
    description: Development server
...
paths:
...
  /calendar/calendars/{calendarId}:
    get:
      tags: [Calendar]
      summary: Get calendar information
      parameters:
        - name: calendarId
          in: path
          required: true
      responses:
        '200':
          description: Calendar information
  /calendar/calendars/{calendarId}/events:
    get:
      tags: [Calendar]
      summary: List events in calendar
      parameters:
        - name: calendarId
          in: path
          required: true
        - name: timeMin
          in: query
          schema:
            type: string
            format: date-time
        - name: timeMax
          in: query
          schema:
            type: string
            format: date-time
        - name: maxResults
          in: query
          schema:
            type: integer
            default: 2500
        - name: singleEvents
          in: query
          schema:
            type: boolean
            default: true
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [startTime, updated]
            default: startTime
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          type: object
                      count:
                        type: integer

    post:
      tags: [Calendar]
      summary: Create event
      parameters:
        - name: calendarId
          in: path
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - startISO
                - endISO
              properties:
                title:
                  type: string
                  minLength: 1
                startISO:
                  type: string
                  format: date-time
                endISO:
                  type: string
                  format: date-time
                description:
                  type: string
                location:
                  type: string
                attendees:
                  type: array
                  items:
                    type: string
                    format: email
                timezone:
                  type: string
                  default: UTC
                sessionId:
                  type: string
      responses:
        '201':
          description: Event created
  /calendar/calendars/{calendarId}/events/{eventId}:
    get:
      tags: [Calendar]
      summary: Get event
      parameters:
        - name: calendarId
          in: path
          required: true
        - name: eventId
          in: path
          required: true
      responses:
        '200':
          description: Event details

    patch:
      tags: [Calendar]
      summary: Update event
      parameters:
        - name: calendarId
          in: path
          required: true
        - name: eventId
          in: path
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                startISO:
                  type: string
                  format: date-time
                endISO:
                  type: string
                  format: date-time
                description:
                  type: string
                location:
                  type: string
                attendees:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Event updated

    delete:
      tags: [Calendar]
      summary: Delete event
      parameters:
        - name: calendarId
          in: path
          required: true
        - name: eventId
          in: path
          required: true
      responses:
        '200':
          description: Event deleted
  /calendar/events:
    post:
      tags: [Calendar]
      summary: Create event (alternative endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - startISO
                - endISO
                - calendarId
              properties:
                title:
                  type: string
                  minLength: 1
                startISO:
                  type: string
                  format: date-time
                endISO:
                  type: string
                  format: date-time
                calendarId:
                  type: string
                description:
                  type: string
                location:
                  type: string
                attendees:
                  type: array
                  items:
                    type: string
                sessionId:
                  type: string
      responses:
        '201':
          description: Event created